Fixed variable types for 64bit compatibility in Base64 decoding.
* src/gsad_omp.c (get_lsc_credentials_omp): Changed result_len to gsize.
(get_report_omp): Changed report_len to gsize.

* src/gsad_omp.h: Updated.

* src/gsad.c: Changed response_size and header_auth_decoded_len to
gsize.


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/gsa@6235 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/gsad.c
@@ -131,7 +131,7 @@ char *content_disposition = NULL;
 /**
  * @brief Response size passed between callbacks when sending files.
  */
-unsigned int response_size = 0;
+gsize response_size = 0;
 
 /**
  * @brief Logging parameters, as passed to setup_log_handlers.
@@ -303,7 +303,7 @@ get_header_credentials (struct MHD_Connection * connection)
   const char *header_auth;
   guchar *header_auth_decoded = NULL;
   const char *strbase = "Basic ";
-  unsigned int header_auth_decoded_len;
+  gsize header_auth_decoded_len;
   gchar **auth_split;
 
   header_auth = MHD_lookup_connection_value (connection,
