+++ src/gsad_omp.h
@@ -51,12 +51,12 @@ char * get_status_omp (credentials_t *, const char *, const char *,
 
 char * delete_report_omp (credentials_t *, const char *, const char *);
 char * get_report_omp (credentials_t *, const char *, const char *,
-                       unsigned int *, const unsigned int,
+                       gsize *, const unsigned int,
                        const unsigned int, const char *, const char *,
                        const char *);
 
 char * get_lsc_credentials_omp (credentials_t *, const char *, const char *,
-                                unsigned int *, const char *, const char *);
+                                gsize *, const char *, const char *);
 char * create_lsc_credential_omp (credentials_t *, char *, char *,
                                   const char *, const char *, const char *);
 char * delete_lsc_credential_omp (credentials_t *, const char *);
